made daterange component
