+++ src/screens/Home/organisms/Events/molecules/index.js
@@ -1,5 +1,4 @@
 import DateBox from './DateBox/DateBox';
-import EventStrip from './EventStrip/EventStrip';
 import EventStripDetails from './EventStripDetails/EventStripDetails';
 
-export { DateBox, EventStrip, EventStripDetails };
+export { DateBox, EventStripDetails };
