+++ src/screens/Home/organisms/Events/organisms/index.js
@@ -1,4 +1,5 @@
 import EventsList from './EventsList/EventsList';
+import EventStrip from './EventStrip/EventStrip';
 import Legend from './Legend/Legend';
 
-export { EventsList, Legend };
+export { EventsList, EventStrip, Legend };
